Bugfixes for file descriptor hash table: - Moved next_fd and hash_table to be static variables in filesys.c - Initialized hash tables before creating the free-map - Moved setting file descriptors/hashing files into filesys_open - Currently will run insult, echo, etc. I believe it also passes create, although lack of process_wait makes this uncertain.
