+++ src/filesys/filesys.h
@@ -20,5 +20,7 @@ bool filesys_create(const char *name, off_t initial_size);
 struct file *filesys_open(const char *name);
 bool filesys_remove(const char *name);
 
+struct file *file_lookup_from_fd(int fd);
+
 #endif /* filesys/filesys.h */
